% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aggregation.r
\name{Aggregation}
\alias{Aggregation}
\title{Aggregation of matrices list.}
\usage{
Aggregation(
  ctrlMatrices.lst = NULL,
  matrices.lst = NULL,
  minDist.num = NULL,
  maxDist.num = NULL,
  trans.fun = NULL,
  agg.fun = "mean",
  rm0.bln = FALSE,
  diff.fun = "substraction",
  scaleCorrection.bln = FALSE,
  correctionArea.lst = NULL,
  statCompare.bln = FALSE
)
}
\arguments{
\item{ctrlMatrices.lst}{<list[matrix]>: The matrices list to aggregate as control.}

\item{matrices.lst}{<list[matrix]>: The matrices list to aggregate.}

\item{minDist.num}{<numeric>: The minimal distance between anchor and bait.}

\item{maxDist.num}{<numeric>: The maximal distance between anchor and bait.}

\item{trans.fun}{<function or chracter>: The function use to transforme or scale values in each submatrix before aggregation. If the parameter is character so:
\itemize{
\item "quantile" or "qtl" apply function dplyr::ntile(x,500)
\item "percentile" or "prct" apply percentile.
\item "rank" apply a ranking.
\item "zscore" apply a scaling.
\item "minmax" apply a Genomic2DTK::MinMaxScale.
\item "mu" apply a Genomic2DTK::MeanScale.
\item other or NULL don't apply transformation (Default).
}}

\item{agg.fun}{<function or chracter>: The function use to aggregate each pixel in matrix. If the parameter is a character so:
\itemize{
\item "50%" or "median" apply the median
\item "+" or "sum" apply the sum
\item other (Default) apply the mean
}}

\item{rm0.bln}{<logical>: if TURE 0 are replace with NA. (Default FALSE)}

\item{diff.fun}{<function or chracter>: The function use to compute differential. If the parameter is character so:
\itemize{
\item "-", "substract" or "substraction" apply a substraction
\item "/" or "ratio" apply a ratio
\item "log2","log2-","log2/" or "log2ratio" apply a log2 on ratio
\item other (Default) apply a log2 on 1+ratio
}}

\item{scaleCorrection.bln}{<logical>: Whether a correction should be done on the median value take in ane noising area. (Default TRUE)}

\item{correctionArea.lst}{<list>: Nested list of indice that define a noising area fore correction. List muste contain in first an element "i" (row indices) then an element called "j" (columns indices). If NULL automatically take in upper left part of aggregated matrices. (Default NULL)}

\item{statCompare.bln}{<logical>: Whether a t.test must be apply to each pxl of the differential aggregated matrix.}
}
\value{
A matrix
}
\description{
Aggregates all the matrices of a list (or two lists in case of differential aggregation) into a single matrix. This function allows to apply different aggregation (average, sum, ...), transformation (rank, percentage, ...) and differential (subtraction, ratio, ...) functions.
}
\details{
Aggregation
}
\examples{
\dontrun{
  # Classical Aggregation
    # submatrixRF_Ctrl.mtx_lst is a submatrix list
    # obtain with Genomic2DTK::ExtractSubmatrix function
  aggreg.mtx <- Aggregation(
    matrices.lst = submatrixRF_Ctrl.mtx_lst, 
    agg.fun      = "sum",
    trans.fun    = "qtl", 
    rm0.bln      = TRUE,
    minDist      = 9000,
    maxDist      = 11000
  )
 
  # Differential Aggregation
    # submatrixRF.mtx_lst is a second submatrix list
    # obtain with Genomic2DTK::ExtractSubmatrix function
  diffAggreg.mtx <- Aggregation(
    ctrlMatrices.lst    = submatrixRF_Ctrl.mtx_lst,
    matrices.lst        = submatrixRF.mtx_lst,
    minDist             = 9000,
    maxDist             = 11000,
    agg.fun             = "mean",
    rm0.bln             = FALSE,
    diff.fun            = "substraction",
    scaleCorrection.bln = TRUE,
    correctionArea.lst  =  list(
      i = c(1:30),
      j = c(72:101)
      ),
    statCompare.bln = TRUE
  )
}
}
