% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiocParallelParam.r
\name{BiocParallelParam}
\alias{BiocParallelParam}
\title{Create parallel parameter}
\usage{
BiocParallelParam(cores.num = 1, verbose.bln = FALSE)
}
\arguments{
\item{cores.num}{<numerical> : An integer to specify the number of cores. (Default 1)}
}
\description{
Create BiocParallel parameter according to OS.
}
\details{
BiocParallelParam
}
\keyword{internal}
