% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CutHiC.r
\name{CutHiC}
\alias{CutHiC}
\title{Cut HiC map in HiC chunks.}
\usage{
CutHiC(megaHic.cmx, cores.num = 1, verbose.bln = TRUE)
}
\arguments{
\item{megaHic.cmx}{<contactMatrix>: The HiC megamap.}

\item{cores.num}{<numerical> : An integer to specify the number of cores. (Default 1)}

\item{verbose.bln}{<logical>: A logical value. If TRUE show the progression in console. (Default TRUE)}
}
\value{
A matrices list.
}
\description{
Cut a mega contactMatrix (join of multiple chromosomic maps) inq a list of contactMatrix.
}
\details{
CutHiC
}
