% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpectedHiC.r
\name{ExpectedHiC}
\alias{ExpectedHiC}
\title{Compute expected.}
\usage{
ExpectedHiC(hic.cmx_lst, verbose.bln = TRUE, cores.num = 1)
}
\arguments{
\item{hic.cmx_lst}{<List[contactMatrix]>: The HiC maps list.}

\item{verbose.bln}{<logical>: A logical value. If TRUE show the progression in console. (Default TRUE)}

\item{cores.num}{<numerical> : An integer to specify the number of cores. (Default 1)}
}
\value{
A matrices list.
}
\description{
Compute the mean of contact as a function of distance and apply as expected value.
}
\details{
ExpectedHiC
}
