% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractSubmatrix.r
\name{ExtractSubmatrix}
\alias{ExtractSubmatrix}
\title{ExtractSubmatrix}
\usage{
ExtractSubmatrix(
  feature.gn = NULL,
  hic.cmx_lst = NULL,
  referencePoint.chr = "rf",
  res.num = NULL,
  matriceDim.num = 21,
  shiftFactor.num = 1,
  cores.num = 1,
  verbose.bln = TRUE
)
}
\arguments{
\item{feature.gn}{<GRanges or Pairs[GRanges] or GInteractions>: The genomic feature on which compute the extraction of HiC submatrix.}

\item{hic.cmx_lst}{<List[contactMatrix]>: The HiC maps list.}

\item{referencePoint.chr}{<character>: A character that give the kind of extraction. "rf" to extract regions or "pf" to extract points interactions. (Default "rf")}

\item{res.num}{<numeric>: the resoulution in used in hic.cmx_lst. If NULL automatically find in resolution attributes of hic.cmx_lst. (Default NULL)}

\item{matriceDim.num}{<numeric>: The size of matrices in output. (Default 21).}

\item{shiftFactor.num}{<numeric>: If "referencePoint.chr" is "rf". How much of the distance between anchor and bait is report before and after the region (Default 1).}

\item{cores.num}{<integer> : An integer to specify the number of cores. (Default 1)}

\item{verbose.bln}{<logical>: A logical value. If TRUE show the progression in console. (Default TRUE)}
}
\value{
A matrices list.
}
\description{
Extract matrices in the HiC maps list around genomic features.
}
\examples{
library(GenomicED)
data("interactions.gni")
data("HiC_ctrl.cmx_lst")


interactions_RFmatrix_ctrl.lst  <- ExtractSubmatrix(
  feature.gn         = interactions.gni,
  hic.cmx_lst        = HiC_ctrl.cmx_lst,
  res.num            = NULL,
  referencePoint.chr = "rf",
  matriceDim.num     = 101,
  cores.num          = 1,
  verbose.bln        = FALSE
  )

str(interactions_RFmatrix_ctrl.lst,max.level = 1, list.len=5)
attributes(interactions_RFmatrix_ctrl.lst)$interactions[1]
}
