% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetQuantif.r
\name{GetQuantif}
\alias{GetQuantif}
\title{GetQuantif}
\usage{
GetQuantif(
  matrices.lst,
  area.fun = "center",
  operation.fun = "mean_rm0",
  name.chr = NULL
)
}
\arguments{
\item{matrices.lst}{<List[matrix]> : A matrices list.}

\item{area.fun}{<character or function> : A character or function thaht alow an extraction of an area on each matrix that compose the matrices lits (Default "center").
\itemize{
\item "mean_rm0" : apply a mean after replace 0 with NA
\item "median_rm0" : apply a median after replace 0 with NA
\item "sum_rm0" : apply a sum after replace 0 with NA
\item "median" : apply a median
\item "sum" : apply a sum
\item "mean" or other character :  apply a mean
}}

\item{operation.fun}{<character or function> : A character or function to applt on the  extracted area of each matrix that compose the matrices lits (Default "mean_rm0").
\itemize{
\item "C" or "CENTER" : Take the area of interaction between anchor and bait.
\item "UL" or "UPPER_LEFT" : Take the area of interaction in the uppper left square
\item "UR" or "UPPER_RIGHT" : Take the area of interaction in the uppper right square
\item "BL" or "BOTTOM_LEFT" : Take the area of interaction in the bottom left square
\item "BR" or "BOTTOM_RIGHT" : Take the area of interaction in the bottom right square
\item "U" or "UPPER" : Take the area of interaction in above the center area
\item "B" or "BOTTOM" : Take the area of interaction in below the center area
\item "L" or "LEFT" : Take the area of interaction in the left of the center area
\item "R" or "RIGHT" : Take the area of interaction in the right of the center area
\item "D" or "DONUT" : Take the area of interaction that surround the center area
}}

\item{name.chr}{<character> : The name of a column in GInteraction attributes of matrices.lst used as named in the output (Default NULL).}
}
\value{
A GRange object.
}
\description{
Compute a function on each matrix that compose a matrices list.
}
