% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GnpToCm.r
\name{GnpToCm}
\alias{GnpToCm}
\title{GnpToCm}
\usage{
GnpToCm(
  hic.gnp = NULL,
  res.num = NULL,
  chromSize.dtf = NULL,
  verbose.bln = TRUE,
  cores.num = 1
)
}
\arguments{
\item{hic.gnp}{<Pairs[GRanges]>: some GRanges or a list of GRanges or a GRangesList}

\item{res.num}{<integer>: A number resolution of the HiC.}

\item{chromSize.dtf}{<data.frame>: A data.frame of genome where first colum correspond to the chromosomes names, and the second column correspond to the chromosomes lengths in base pairs.}

\item{verbose.bln}{<logical>: A logical value. If TRUE show the progression in console. (Default TRUE)}

\item{cores.num}{<integer> : An integer to specify the number of cores. (Default 1)}
}
\value{
A list of ContactMatrix.
}
\description{
Convert a Pairs of GRanges in an list of ContactMatrix (matricies are stored in sparseMatrix format).
}
