% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hsl2Rgb.r
\name{Hsl2Rgb}
\alias{Hsl2Rgb}
\title{Convert a color in HSL to RGB.}
\usage{
Hsl2Rgb(hsl.col = NULL, alpha.bln = FALSE)
}
\arguments{
\item{hsl.col}{<charcater>: a vector of the color's HSL code.}

\item{alpha.bln}{<logical>: whether the alpha layer should be returned. (Default FALSE)}
}
\value{
An integer vector of the color's RGB code.
}
\description{
Convert a color in HSl (Hue,Saturation,Light) format to RGB format.
}
\details{
Hsl2Rgb
}
\examples{
Hsl2Rgb(c(43.8,0.873,0.492,0.498),alpha.bln=TRUE)
}
\keyword{internal}
