% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICEnorm.r
\name{ICEnorm}
\alias{ICEnorm}
\title{Compute Iterative Correction on hic chunk.}
\usage{
ICEnorm(hic.cmx, qtlTh.num = 0.15, maxIter.num = 50)
}
\arguments{
\item{hic.cmx}{<contactMatrix>: The HiC maps chunk to normalize.}

\item{qtlTh.num}{<numerical>: the threshold quantile below which the bins will be ignored. (Default 0.15)}

\item{maxIter.num}{<numerical>: The maximum iteration number.}
}
\value{
A normalized contactMatrix
}
\description{
Compute Iterative Correction (Vanilla Count) on hic maps.
}
\details{
ICEnorm
}
