% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndexFeatures.r
\name{IndexFeatures}
\alias{IndexFeatures}
\title{IndexFeatures}
\usage{
IndexFeatures(
  gRange.gnr_lst = NULL,
  constraint.gnr = NULL,
  chromSize.dtf = NULL,
  binSize.int = NULL,
  method.chr = "mean",
  variablesName.chr_vec = NULL,
  cores.num = 1,
  verbose.bln = TRUE
)
}
\arguments{
\item{gRange.gnr_lst}{<GRanges or GRangesList or list[GRanges]>: some GRanges or a list of GRanges or a GRangesList}

\item{constraint.gnr}{<GRanges>: A GRange of constraint regions. If NULL chromosomes in chromSize.dtf are used (Default NULL)}

\item{chromSize.dtf}{<data.frame>: A data.frame of genome where first colum correspond to the chromosomes names, and the second column correspond to the chromosomes lengths in base pairs.}

\item{binSize.int}{<integer>: A number that specify the width bins.}

\item{method.chr}{<character>: A string of a summary method name as 'mean', 'median', 'sum', 'max, 'min'. (Default 'mean'')}

\item{variablesName.chr_vec}{<character> : A character vector that specify the metadata columns of GRanges on which apply the summary method.}

\item{cores.num}{<integer> : An integer to specify the number of cores. (Default 1)}

\item{verbose.bln}{<logical>: A logical value. If TRUE show the progression in console. (Default TRUE)}
}
\value{
A GRange object.
}
\description{
Bin multiple GRanges and summary all in one GRange. Could overlap ranges with constraints regions
}
\examples{
library(GenomicED)
data("anchors_Peaks.gnr")
anchors_Peaks.gnr[1]


seqlengths.num <- c('2L'=23513712, '2R'=25286936)
chromSize.dtf  <- data.frame(
  seqnames   = names(seqlengths.num ), 
  seqlengths = seqlengths.num
  )
binSize.num <- 10000

anchors_Index.gnr <- IndexFeatures(
  gRange.gnr_lst        = list(Beaf=anchors_Peaks.gnr), 
  constraint.gnr        = domains.gnr,
  chromSize.dtf         = chromSize.dtf,
  binSize.int           = binSize.num,
  method.chr            = "max",
  variablesName.chr_vec = "score",
  cores.num             = 1,
  verbose.bln           = FALSE
  )

anchors_Index.gnr[1]
}
