% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JoinHiC.r
\name{JoinHiC}
\alias{JoinHiC}
\title{Merge HiC chunk as HiC map}
\usage{
JoinHiC(hic.cmx_lst)
}
\arguments{
\item{hic.cmx_lst}{<List[contactMatrix]>: The HiC maps list.}
}
\value{
A contactMatrix.
}
\description{
Create mega contactMatrix from a list of contactMatrix.
}
\details{
JoinHiC
}
