% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrienteMatrix.r
\name{OrienteMatrix}
\alias{OrienteMatrix}
\title{OrienteMatrix}
\usage{
OrienteMatrix(matrice.mtx)
}
\arguments{
\item{matrice.mtx}{<matrix or List[matrix]>: matrix or matricies list to oriente}
}
\value{
Oriented matrix or matricies list
}
\description{
Apply a 180° rotation follow with a transposation on a matrix or on matricies in a list according to the interactions attributes of the list.
}
\examples{
library(GenomicED)
data("submatrixPF_Ctrl.mtx_lst")


head(attributes(submatrixPF_Ctrl.mtx_lst)$interactions$orientation)

submatrixPF_Ctrl_oriented.mtx_lst <- OrienteMatrix(submatrixPF_Ctrl.mtx_lst)

head(attributes(submatrixPF_Ctrl_oriented.mtx_lst)$interactions$orientation)
}
