% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotAPA.r
\name{PlotAPA}
\alias{PlotAPA}
\title{Draw aggregation plot.}
\usage{
PlotAPA(
  apa.mtx = NULL,
  trimPrct.num = 0,
  colMin.num = NULL,
  colMid.num = NULL,
  colMax.num = NULL,
  colCondMin.num = NULL,
  colCondMax.num = NULL
)
}
\arguments{
\item{apa.mtx}{<matrix>: The aggregated matrix.}

\item{trimPrct.num}{<numeric>: A number between 0 and 100 thaht give the percentage of triming in matrices.}

\item{colMin.num}{<matrix>: The minimal value in color scale. If Null automaticaly find.}

\item{colMid.num}{<matrix>: The middle value in color scale. If Null automaticaly find.}

\item{colMax.num}{<matrix>: The mximal value in color scale. If Null automaticaly find.}

\item{colCondMin.num}{<matrix>: Avalaible for plotting differantial aggregation. The minimal value in color scale in the classsical aggregation plot. If Null automaticaly find.}

\item{colCondMax.num}{<matrix>: Avalaible for plotting differantial aggregation. The maxiaml value in color scale in the classsical aggregation plot. If Null automaticaly find.}
}
\value{
None
}
\description{
Draw aggregation plot from aggregation matrices.
}
\details{
PlotAPA
}
\examples{
library(GenomicED)
data(aggreg.mtx)


PlotAPA(
    apa.mtx        = aggreg.mtx,
    trimPrct.num   = 20,
    colMin.num     = -2,
    colMid.num     = 0,
    colMax.num     = 2,
    colCondMin.num = 0,
    colCondMax.num = 2
)
}
