% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SearchPairs.r
\name{SearchPairs}
\alias{SearchPairs}
\title{Creates pairs of coordinates from indexed anchor and bait genomic coordinates according to distance constraints.}
\usage{
SearchPairs(
  indexAnchor.gnr = NULL,
  indexBait.gnr = NULL,
  minDist.num = NULL,
  maxDist.num = NULL,
  verbose.bln = TRUE,
  cores.num = 1
)
}
\arguments{
\item{indexAnchor.gnr}{<GRanges>: A first indexed GRanges object used as pairs anchor (must be indexed using IndexFeatures()).}

\item{indexBait.gnr}{<GRanges>: A second indexed GRanges object used as pairs bait (must be indexed using IndexFeatures()). If NULL, indexAnchor.gnr is used instead (Default NULL)}

\item{minDist.num}{<numeric>: Minimal distance between anchors and baits. (Default NULL)}

\item{maxDist.num}{<numeric>: Maximal distance between anchors and baits. (Default NULL)}

\item{verbose.bln}{<logical>: If TRUE show the progression in console. (Default TRUE)}

\item{cores.num}{<integer> : Number of cores to use. (Default 1)}
}
\value{
A GInteractions object.
}
\description{
Creates pairs of coordinates from indexed anchor and bait genomic coordinates according to distance constraints.
}
\details{
SearchPairs
}
\examples{
library(GenomicED)

data(anchors_Index.gnr)
anchors_Index.gnr[1]

data(baits_Index.gnr)
baits_Index.gnr[1]


interactions.gni <- SearchPairs(
    indexAnchor.gnr = anchors_Index.gnr,
    indexBait.gnr   = baits_Index.gnr,
    minDist.num     = 9000, 
    maxDist.num     = 11000,
    cores.num       = 1,
    verbose.bln     = FALSE
    )

interactions.gni[1]
}
