% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ShowLoading}
\alias{ShowLoading}
\title{Show the loading of a looping operation.}
\usage{
ShowLoading(start.tim = NULL, operation.ndx = NULL, operation.num = NULL)
}
\arguments{
\item{start.tim}{<POSIXct POSIXt>: A time obtain with 'Sys.time()'.}

\item{operation.ndx}{<numerical>: The index number of the operation.}

\item{operation.num}{<numerical>: The total number of operation.}
}
\description{
Show the loading of a looping operation.
}
\details{
ShowLoading
}
\examples{
start.tim <- Sys.time()
for(i in seq_len(10000)){
    ShowLoading(start.tim, i , 10000)
}
}
