% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SwitchMatrix.r
\name{SwitchMatrix}
\alias{SwitchMatrix}
\title{Change values in hic chunk}
\usage{
SwitchMatrix(hic.cmx_lst, matrixKind.chr = c("obs", "norm", "o/e", "exp"))
}
\arguments{
\item{hic.cmx_lst}{<List[contactMatrix]>: The HiC maps list.}

\item{matrixKind.chr}{<character>: The kind of matrix you want.}
}
\value{
A contactMatrix list.
}
\description{
Change values in matrix with observed, balanced, observed/expected or expected values according to what are be done in hic.
}
\details{
SwitchMatrix
}
