% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCnorm.r
\name{VCnorm}
\alias{VCnorm}
\title{Compute Vanilla Count or Vanilla Count square root correction on hic chunk.}
\usage{
VCnorm(hic.cmx = NULL, qtlTh.num = 0.15, sqrt.bln = TRUE)
}
\arguments{
\item{hic.cmx}{<contactMatrix>: The HiC maps chunk to normalize.}

\item{qtlTh.num}{<numerical>: the threshold quantile below which the bins will be ignored. (Default 0.15)}

\item{sqrt.bln}{<logical>: Whether the square root must be apply. (Default TRUE)}
}
\value{
A matrices list.
}
\description{
Compute Vanilla Count or Vanilla Count square root correction normalization on hic maps.
}
\details{
VCnorm
}
