% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YlGnBu.R
\name{YlGnBu}
\alias{YlGnBu}
\title{YlGnBu palette.}
\usage{
YlGnBu(
  paletteLength.num = NULL,
  space.chr = "rgb",
  interpolate.chr = "linear",
  bias.num = 1
)
}
\arguments{
\item{paletteLength.num}{<numeric>: color number.}

\item{space.chr}{<numeric>: a character string; interpolation in RGB or CIE Lab color spaces. See ?grDevices::colorRamp for more details. (Default "rgb")}

\item{interpolate.chr}{<numeric>: use spline or linear interpolation. See ?grDevices::colorRamp for more details. (Default "linear")}

\item{bias.num}{<numeric>: a positive number. Higher values give more widely spaced colors at the high end. See ?grDevices::colorRamp for more details. (Default 1)}
}
\value{
A vector of color.
}
\description{
Create a YlGnBu palette.
}
\details{
YlGnBu
}
\examples{
YlGnBu(9)

}
