% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpectedHiC.R
\name{ExpectedHiC}
\alias{ExpectedHiC}
\title{Genomic distance bias correction.}
\usage{
ExpectedHiC(hic.cmx_lst, verbose.bln = FALSE, cores.num = 1)
}
\arguments{
\item{hic.cmx_lst}{<List[contactMatrix]>: The HiC maps list.}

\item{verbose.bln}{<logical>: If TRUE show the progression in console. (Default FALSE)}

\item{cores.num}{<numerical> : Number of cores to be used. (Default 1)}
}
\value{
A matrices list.
}
\description{
Function that normalises HiC matrices by expected values computing by genomic distance.
}
\details{
ExpectedHiC
}
\examples{
# Note: run Genomic2DTK::NormalizeHiC before ExpectedHiC calculation.
data(HiC_Ctrl.cmx_lst)
ExpectedHiC(HiC_Ctrl.cmx_lst)
}
