% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalizeHiC.r
\name{NormalizeHiC}
\alias{NormalizeHiC}
\title{Compute HiC map matrix-balancing normalization.}
\usage{
NormalizeHiC(
  hic.cmx_lst,
  method.chr = "ICE",
  interaction.type = NULL,
  maxIter.num = 50,
  qtlTh.num = 0.15,
  cores.num = 1,
  verbose.bln = TRUE
)
}
\arguments{
\item{hic.cmx_lst}{<List[contactMatrix]>: The HiC maps list.}

\item{method.chr}{<character> : The kind of normalization method. One of "ICE", "VC" or "VC_SQRT" (Default "ICE")}

\item{interaction.type}{<character> : "cis", "trans", c("cis", "trans"), "all". If NULL normalization is apply on cis chunks then trans chunks (equivalent to c("cis", "trans")). If is "all", normalization is apply on all chuncks at once. (Default NULL)}

\item{maxIter.num}{<numerical>: The maximum iteration number.}

\item{qtlTh.num}{<numerical>: the threshold quantile below which the bins will be ignored. (Default 0.15)}

\item{cores.num}{<numerical> : Number of cores to be used. (Default 1)}

\item{verbose.bln}{<logical>: A logical value. If TRUE show the progression in console. (Default TRUE)}
}
\value{
A matrices list.
}
\description{
Apply a matrix-balancing normalization method to a list of contacts matrix.
}
\details{
NormalizeHiC
}
\examples{
HiC_Ctrl_ICE.cmx_lst     <- NormalizeHiC(HiC_Ctrl.cmx_lst,
    interaction.type="cis",
    method.chr="ICE")
    
HiC_Ctrl_VC.cmx_lst      <- NormalizeHiC(HiC_Ctrl.cmx_lst,
    interaction.type=c("cis","trans"),
    method.chr="VC")

HiC_Ctrl_VC_SQRT.cmx_lst <- NormalizeHiC(HiC_Ctrl.cmx_lst,
    interaction.type="all",
    method.chr="VC_SQRT")
}
