% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalizeHiC.r
\name{NormalizeHiC}
\alias{NormalizeHiC}
\title{Comoute HiC map normalization.}
\usage{
NormalizeHiC(
  hic.cmx_lst,
  method.chr = "ICE",
  type.chr = c("cis", "trans", "all"),
  maxIter.num = 50,
  qtlTh.num = 0.15,
  cores.num = 1,
  verbose.bln = TRUE
)
}
\arguments{
\item{hic.cmx_lst}{<List[contactMatrix]>: The HiC maps list.}

\item{method.chr}{<character> : The kind of normalization method. One of "ICE", "VC" or "VC_SQRT" (Default "ICE")}

\item{type.chr}{<character> : Whether normalization must be apply on "cis" chunks, "trans" chunks or "all" chuncks. (Default "cis")}

\item{maxIter.num}{<numerical>: The maximum iteration number.}

\item{qtlTh.num}{<numerical>: the threshold quantile below which the bins will be ignored. (Default 0.15)}

\item{cores.num}{<numerical> : An integer to specify the number of cores. (Default 1)}

\item{verbose.bln}{<logical>: A logical value. If TRUE show the progression in console. (Default TRUE)}
}
\value{
A matrices list.
}
\description{
Apply a normalization method to a list of contacts matrix.
}
\details{
NormalizeHiC
}
