% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotAPA.r
\name{PlotAPA}
\alias{PlotAPA}
\title{PlotAPA}
\usage{
PlotAPA(
  apa.mtx = NULL,
  trimPrct.num = 0,
  minBoundary.num = NULL,
  center.num = NULL,
  maxBoundary.num = NULL,
  minConditionBoundary.num = NULL,
  maxConditionBoundary.num = NULL
)
}
\arguments{
\item{apa.mtx}{<matrix>: The aggregated matrix.}

\item{trimPrct.num}{<numeric>: A number between 0 and 100 thaht give the percentage of triming in matrices.}

\item{minBoundary.num}{<matrix>: The minimal value in color scale. If Null automaticaly find.}

\item{center.num}{<matrix>: The middle value in color scale. If Null automaticaly find.}

\item{maxBoundary.num}{<matrix>: The mximal value in color scale. If Null automaticaly find.}

\item{minConditionBoundary.num}{<matrix>: Avalaible for plotting differantial aggregation. The minimal value in color scale in the classsical aggregation plot. If Null automaticaly find.}

\item{maxConditionBoundary.num}{<matrix>: Avalaible for plotting differantial aggregation. The maxiaml value in color scale in the classsical aggregation plot. If Null automaticaly find.}
}
\value{
None
}
\description{
Draw aggregation plot from aggregation matrices.
}
\examples{
library(GenomicED)
data(aggreg.mtx)
PlotAPA(
    apa.mtx                  = aggreg.mtx,
    trimPrct.num             = 20,
    minBoundary.num          = -2,
    center.num               = 0,
    maxBoundary.num          = 2,
    minConditionBoundary.num = 0,
    maxConditionBoundary.num = 2
)
}
